+++ NEWS.md
@@ -5,6 +5,7 @@
 
 #### Highlights
 * Update ReadDescription() to read information about point source types
+* Allow adding comment line to WriteObs()
 
 ## HYPEtools 1.6.3 (2024-09-04)
 *Enhancements and bug fixes*